requirements.txt
streamlit
opencv-python
transformers
torch
pillow
cmake
dlib
face_recognition
